+++ src/main/java/org/zerock/controller/HomeController.java
@@ -36,7 +36,7 @@ public class HomeController {
 		
 		model.addAttribute("serverTime", formattedDate );
 		
-		return "element";
+		return "board/read";
 	}
 	
 }
