+++ src/main/java/org/zerock/domain/Criteria.java
@@ -1,5 +1,11 @@
 package org.zerock.domain;
 
+import java.io.UnsupportedEncodingException;
+import java.net.URLEncoder;
+
+import org.springframework.web.util.UriComponents;
+import org.springframework.web.util.UriComponentsBuilder;
+
 import lombok.Data;
 
 @Data
@@ -28,4 +34,32 @@ public class Criteria {
 		}
 		return this.type.split("");
 	}
+	
+	public String makeSearch(int page) {
+        if(this.keyword!=null || this.type!=null) {
+        
+        UriComponents uriComponents =
+                UriComponentsBuilder.newInstance().queryParam("page", page).
+                queryParam("type", this.type).
+                queryParam("keyword", encoding(this.keyword)).build();
+        return uriComponents.toUriString();}
+        
+        else {UriComponents uriComponents =
+                UriComponentsBuilder.newInstance().queryParam("page", page).
+                build();
+        return uriComponents.toUriString();}
+                    
+    }
+	
+    
+    private String encoding(String keyword) {
+        if(keyword == null || keyword.trim().length()==0) {
+            return "";
+        }
+        try {
+            return URLEncoder.encode(keyword,"UTF-8");
+        }catch(UnsupportedEncodingException e) {
+            return "";
+        }
+    }
 }
