+++ src/main/java/org/zerock/domain/PageMaker.java
@@ -1,7 +1,6 @@
 package org.zerock.domain;
 
-import java.io.UnsupportedEncodingException;
-import java.net.URLEncoder;
+
 
 import org.springframework.web.util.UriComponents;
 import org.springframework.web.util.UriComponentsBuilder;
@@ -35,41 +34,10 @@ public class PageMaker {
 		}
 		this.end = tempLast;
 	}
-	
+
 	public String makeQuery(int page) {
-		 UriComponents uriComponents =
-	                UriComponentsBuilder.newInstance().queryParam("page", page).
-	                build();
-	                return uriComponents.toUriString();
-	                }
-		
-	
-	
-	public String makeSearch(int page) {
-        if(cri.getKeyword()!=null || cri.getType()!=null) {
-        
-        UriComponents uriComponents =
-                UriComponentsBuilder.newInstance().queryParam("page", page).
-                queryParam("type", cri.getType()).
-                queryParam("keyword", encoding(cri.getKeyword())).build();
-        return uriComponents.toUriString();}
-        
-        else {UriComponents uriComponents =
-                UriComponentsBuilder.newInstance().queryParam("page", page).
-                build();
-        return uriComponents.toUriString();}
-                    
-    }
-    
-    private String encoding(String keyword) {
-        if(keyword == null || keyword.trim().length()==0) {
-            return "";
-        }
-        try {
-            return URLEncoder.encode(keyword,"UTF-8");
-        }catch(UnsupportedEncodingException e) {
-            return "";
-        }
-    }
+		UriComponents uriComponents = UriComponentsBuilder.newInstance().queryParam("page", page).build();
+		return uriComponents.toUriString();
+	}
 
 }
