+++ src/main/java/org/zerock/mapper/BoardMapper.java
@@ -16,5 +16,5 @@ public interface BoardMapper {
 	
 	public int register(BoardVO vo);
 			
-	
+	public BoardVO read(int bno);
 }
