+++ src/main/resources/org/zerock/mapper/BoardMapper.xml
@@ -7,15 +7,16 @@
 
 
 	<sql id="search">
-	<where>
-	<if test="arr!= null">
-	<foreach collection="arr" item="target" open="(" close=")" separator="OR">
-	<if test="target eq 't'.toString()"> title like concat('%',#{keyword},'%') </if>
-	<if test="target eq 'c'.toString()"> content like concat('%',#{keyword},'%') </if>
-	<if test="target eq 'w'.toString()"> writer like concat('%',#{keyword},'%') </if>
-	</foreach>
-	</if>
-	</where>
+		<where>
+			<if test="arr!= null">
+				<foreach collection="arr" item="target" open="(" close=")"
+					separator="OR">
+					<if test="target eq 't'.toString()"> title like concat('%',#{keyword},'%') </if>
+					<if test="target eq 'c'.toString()"> content like concat('%',#{keyword},'%') </if>
+					<if test="target eq 'w'.toString()"> writer like concat('%',#{keyword},'%') </if>
+				</foreach>
+			</if>
+		</where>
 	</sql>
 
 
@@ -35,5 +36,13 @@
 	<insert id="register">
 	insert into tbl_aboard (title,content,writer) values (#{title},#{content},#{writer}); 
 	</insert>
+
+	<select id="read" resultType="org.zerock.domain.BoardVO">
+
+		select title,content,writer,regdate from tbl_aboard where bno=#{bno};
+	</select>
+
+
+
 
 </mapper>
